+++ tst/components/ui/Modal/ModalButtonGroup.test.tsx
@@ -16,7 +16,7 @@ describe('ModalButtonGroup', () => {
   const secondaryButtonLbl = 'test-secondary';
   const onClose = jest.fn();
   const labelID = 'test-label';
-  const testContext = { onClose, labelID };
+  const testContext = { onClose, labelID, dismissible: true, };
 
   it('renders a group with a primary Button', () => {
     const component = (
@@ -110,5 +110,25 @@ describe('ModalButtonGroup', () => {
     fireEvent.click(modalButton);
     expect(component.props.value.onClose).toHaveBeenCalled();
     expect(getByLabelText('close')).toBeInTheDocument();
+  });
+
+  it('does not inject the onClose function into the ModalButton click handler when "dismissible" is false', () => {
+    const onClose = jest.fn();
+    const labelID = 'test-label';
+    const context = { onClose, labelID, dismissible: false };
+    const component = (
+      <ModalContext.Provider value={context}>
+        <ModalButtonGroup
+          primaryButtons={[<ModalButton label="close" closesModal />]}
+        />
+      </ModalContext.Provider>
+    );
+    const { getByTestId } = renderWithTheme(
+      lightTheme,
+      component
+    );
+    const modalButton = getByTestId('button');
+    fireEvent.click(modalButton);
+    expect(component.props.value.onClose).not.toHaveBeenCalled();
   });
 });