+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added poorConnection property to DeskPhone icon
 - Added optional 'id' prop for ui components
 - Added a classname to PopOverMenu component for styling access
+- Added a "dismissible" prop to Modal to optionally allow persistent modals
 
 ### Changed